Added Map functionality

+++ init/index.js
@@ -4,10 +4,14 @@ const mongoose = require("mongoose");
 const initData = require("./data.js");
 const Listing = require("../models/listing.js");
 
-const MONGO_URL = `mongodb://${process.env.MONGODB_URI}`;
+const MONGO_URL = process.env.MONGO_URI;
+
+if (!MONGO_URL) {
+  throw new Error("MONGO_URI is missing in .env");
+}
 
 const OWNER_ID = new mongoose.Types.ObjectId(
-  `${process.env.ADMIN_OBJECT_ID}`
+  process.env.ADMIN_OBJECT_ID
 );
 
 main()
@@ -20,17 +24,26 @@ async function main() {
 
 const initDB = async () => {
   try {
+
+    if (process.env.NODE_ENV !== "development") {
+      throw new Error("Seeding blocked outside development mode");
+    }
+
     await Listing.deleteMany({});
 
     const normalizedData = initData.data.map((item) => ({
       ...item,
       owner: OWNER_ID,
-      image: item.image   // ✅ KEEP OBJECT
+      image: {
+        url: item.image.url,
+        filename: item.image.filename,
+      },
     }));
 
     await Listing.insertMany(normalizedData);
 
     console.log("Database seeded successfully");
+
   } catch (err) {
     console.error("Seeding failed:", err);
   } finally {
